Add more error handling and simplify the usage patterns
