+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/StreamListenerWithAnnotatedArgsTests.java
@@ -16,6 +16,7 @@
 
 package org.springframework.cloud.stream.config;
 
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import org.junit.Test;
@@ -32,20 +33,54 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.MessageHandler;
+import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.MessagingException;
 import org.springframework.messaging.SubscribableChannel;
+import org.springframework.messaging.handler.annotation.Header;
+import org.springframework.messaging.handler.annotation.Headers;
 import org.springframework.messaging.support.MessageBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.fail;
 
 /**
  * @author Marius Bogoevici
+ * @author Ilayaperumal Gopinathan
  */
 public class StreamListenerWithAnnotatedArgsTests {
 
 	@Test
 	public void testInputOutputArgs() throws Exception {
 		ConfigurableApplicationContext context = SpringApplication.run(TestInputOutputArgs.class, "--server.port=0");
+		sendMessageAndValidate(context);
+	}
+
+	@Test
+	public void testInputOutputArgsWithMoreParameters() throws Exception {
+		try {
+			SpringApplication.run(TestInputOutputArgsWithMoreParameters.class, "--server.port=0");
+			fail("Declarative StreamListener method should throw IllegalStateException for invalid method parameters");
+		}
+		catch (Exception e) {
+			assertThat(e.getMessage()).contains("Declarative StreamListener method should only have " +
+					"inbound or outbound targets as method parameters");
+		}
+	}
+
+	@Test
+	public void testInputOutputArgsWithInvalidBindableTarget() throws Exception {
+		try {
+			SpringApplication.run(TestInputOutputArgsWithInvalidBindableTarget.class, "--server.port=0");
+			fail("StreamListener method should throw IllegalStateException for invalid bindable target as method parameter");
+		}
+		catch (Exception e) {
+			assertThat(e.getMessage()).contains("Bindable target is not found for the name:");
+		}
+	}
+
+	@Test
+	public void testInputOutputArgsWithParameterOrderChanged() throws Exception {
+		ConfigurableApplicationContext context = SpringApplication.run(TestInputOutputArgsWithParameterOrderChanged.class, "--server.port=0");
 		sendMessageAndValidate(context);
 	}
 
@@ -67,6 +102,53 @@ public class StreamListenerWithAnnotatedArgsTests {
 
 		@StreamListener
 		public void receive(@Input(Processor.INPUT) SubscribableChannel input, @Output(Processor.OUTPUT) final MessageChannel output) {
+			input.subscribe(new MessageHandler() {
+				@Override
+				public void handleMessage(Message<?> message) throws MessagingException {
+					output.send(MessageBuilder.withPayload(message.getPayload().toString().toUpperCase()).build());
+				}
+			});
+		}
+	}
+
+	@EnableBinding(Processor.class)
+	@EnableAutoConfiguration
+	public static class TestInputOutputArgsWithMoreParameters {
+
+		@StreamListener
+		public void receive(@Input(Processor.INPUT) SubscribableChannel input, @Output(Processor.OUTPUT) final MessageChannel output,
+				@Headers final Map<String, Object> headers,
+				@Header(MessageHeaders.CONTENT_TYPE) final String contentType) {
+			input.subscribe(new MessageHandler() {
+				@Override
+				public void handleMessage(Message<?> message) throws MessagingException {
+					output.send(MessageBuilder.withPayload(message.getPayload().toString().toUpperCase()).build());
+				}
+			});
+		}
+	}
+
+	@EnableBinding(Processor.class)
+	@EnableAutoConfiguration
+	public static class TestInputOutputArgsWithInvalidBindableTarget {
+
+		@StreamListener
+		public void receive(@Input("invalid") SubscribableChannel input, @Output(Processor.OUTPUT) final MessageChannel output) {
+			input.subscribe(new MessageHandler() {
+				@Override
+				public void handleMessage(Message<?> message) throws MessagingException {
+					output.send(MessageBuilder.withPayload(message.getPayload().toString().toUpperCase()).build());
+				}
+			});
+		}
+	}
+
+	@EnableBinding(Processor.class)
+	@EnableAutoConfiguration
+	public static class TestInputOutputArgsWithParameterOrderChanged {
+
+		@StreamListener
+		public void receive(@Output(Processor.OUTPUT) final MessageChannel output, @Input("input") SubscribableChannel input) {
 			input.subscribe(new MessageHandler() {
 				@Override
 				public void handleMessage(Message<?> message) throws MessagingException {
